app/backend/scraper/requirements.txt
attrs==25.3.0
automat==24.8.1
aws-lambda-typing==2.20.0
boto3==1.37.13
boto3-stubs==1.37.13
botocore==1.37.13
botocore-stubs==1.37.13
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.2
cssselect==1.3.0
defusedxml==0.7.1
filelock==3.18.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.3.1
mypy-boto3-s3==1.37.0
packaging==24.2
parsel==1.10.0
pip==24.3.1
protego==0.4.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydispatcher==2.0.7
pyopenssl==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
s3transfer==0.11.4
scrapy==2.12.0
service-identity==24.2.0
setuptools==76.0.0
six==1.17.0
tldextract==5.1.3
twisted==24.11.0
types-awscrt==0.24.2
types-s3transfer==0.11.4
typing-extensions==4.12.2
urllib3==2.3.0
w3lib==2.3.1
zope-interface==7.2
